pyproject.toml
[project]
name = "coding-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.2",
    "google-generativeai>=0.8.5",
    "langchain>=1.2.0",
    "langchain-google-genai>=4.0.0",
    "langgraph>=1.0.5",
    "openai>=2.12.0",
]
